requirements.txt
python-dotenv
matplotlib
pandas
numpy
aiogram==3.0.0
pip install "aiogram>=3.2.0"
